refactor: remove nested transactions directory
- Move models from transaction.transactions to transaction package
- Move repository classes to transaction package
- Move service classes to transaction package
- Update all import statements and package declarations
- Clean up package structure for better organization
