+++ app/transaction-service/src/main/java/com/corporatebanking/transaction/models/CreateTransactionAccountTypeData.java
@@ -1,4 +1,4 @@
-package com.corporatebanking.transaction.transactions.models;
+package com.corporatebanking.transaction.models;
 
 public record CreateTransactionAccountTypeData(
         Long id,
