+++ app/transaction-service/src/main/java/com/corporatebanking/transaction/repository/jdbc/CreateTransactionJdbcRepository.java
@@ -1,0 +1,7 @@
+package com.corporatebanking.transaction.repository.jdbc;
+
+import com.corporatebanking.transaction.models.CreateTransactionData;
+
+public interface CreateTransactionJdbcRepository {
+    CreateTransactionData save(CreateTransactionData createTransactionData);
+}
