+++ app/transaction-service/src/main/java/com/corporatebanking/transaction/repository/jpa/CreateTransactionJpaRepository.java
@@ -1,0 +1,4 @@
+package com.corporatebanking.transaction.repository.jpa;
+
+public interface CreateTransactionJpaRepository {
+}
