--- app/transaction-service/src/main/java/com/corporatebanking/transaction/transactions/repository/jdbc/CreateTransactionJdbcRepository.java
@@ -1,7 +1,0 @@
-package com.corporatebanking.transaction.transactions.repository.jdbc;
-
-import com.corporatebanking.transaction.transactions.models.CreateTransactionData;
-
-public interface CreateTransactionJdbcRepository {
-    CreateTransactionData save(CreateTransactionData createTransactionData);
-}
