--- app/transaction-service/src/main/java/com/corporatebanking/transaction/transactions/repository/jpa/CreateTransactionJpaRepository.java
@@ -1,4 +1,0 @@
-package com.corporatebanking.transaction.transactions.repository.jpa;
-
-public interface CreateTransactionJpaRepository {
-}
